[MZUR-223] Move items into order domain, add gate, add order policy
