+++ database/factories/BarcodeFactory.php
@@ -10,11 +10,6 @@ class BarcodeFactory extends Factory
 {
     protected $model = Barcode::class;
 
-    /**
-     * Define the model's default state.
-     *
-     * @return array<string, mixed>
-     */
     public function definition(): array
     {
         return [
