+++ database/factories/CompanyFactory.php
@@ -5,18 +5,10 @@ namespace Database\Factories;
 use Domain\Companies\Models\Company;
 use Illuminate\Database\Eloquent\Factories\Factory;
 
-/**
- * @extends \Illuminate\Database\Eloquent\Factories\Factory<\Domain\Companies\Models\Company>
- */
 class CompanyFactory extends Factory
 {
     protected $model = Company::class;
 
-    /**
-     * Define the model's default state.
-     *
-     * @return array<string, mixed>
-     */
     public function definition(): array
     {
         return [
