+++ database/factories/ItemFactory.php
@@ -3,21 +3,13 @@
 namespace Database\Factories;
 
 use Domain\Companies\Models\Company;
-use Domain\Items\Models\Item;
+use Domain\Orders\Models\Item;
 use Illuminate\Database\Eloquent\Factories\Factory;
 
-/**
- * @extends \Illuminate\Database\Eloquent\Factories\Factory<\Domain\Items\Models\Item>
- */
 class ItemFactory extends Factory
 {
     protected $model = Item::class;
 
-    /**
-     * Define the model's default state.
-     *
-     * @return array<string, mixed>
-     */
     public function definition(): array
     {
         return [
