+++ database/factories/ProcessFactory.php
@@ -7,18 +7,10 @@ use Domain\Processes\Models\Process;
 use Domain\Statuses\Models\Status;
 use Illuminate\Database\Eloquent\Factories\Factory;
 
-/**
- * @extends \Illuminate\Database\Eloquent\Factories\Factory<\Domain\Processes\Models\Process>
- */
 class ProcessFactory extends Factory
 {
     protected $model = Process::class;
 
-    /**
-     * Define the model's default state.
-     *
-     * @return array<string, mixed>
-     */
     public function definition(): array
     {
         return [
