+++ database/factories/StatusFactory.php
@@ -6,18 +6,10 @@ use Domain\Companies\Models\Company;
 use Domain\Statuses\Models\Status;
 use Illuminate\Database\Eloquent\Factories\Factory;
 
-/**
- * @extends \Illuminate\Database\Eloquent\Factories\Factory<\Domain\Statuses\Models\Status>
- */
 class StatusFactory extends Factory
 {
     protected $model = Status::class;
 
-    /**
-     * Define the model's default state.
-     *
-     * @return array<string, mixed>
-     */
     public function definition(): array
     {
         return [
