+++ database/factories/SubscriptionFactory.php
@@ -2,19 +2,10 @@
 
 namespace Database\Factories;
 
-use Domain\Subscriptions\Models\Subscription;
 use Illuminate\Database\Eloquent\Factories\Factory;
 
-/**
- * @extends Factory<Subscription>
- */
 class SubscriptionFactory extends Factory
 {
-    /**
-     * Define the model's default state.
-     *
-     * @return array<string, mixed>
-     */
     public function definition(): array
     {
         return [
