+++ database/factories/WarehouseFactory.php
@@ -6,18 +6,10 @@ use Domain\Companies\Models\Company;
 use Domain\Warehouses\Models\Warehouse;
 use Illuminate\Database\Eloquent\Factories\Factory;
 
-/**
- * @extends \Illuminate\Database\Eloquent\Factories\Factory<\Domain\Warehouses\Models\Warehouse>
- */
 class WarehouseFactory extends Factory
 {
     protected $model = Warehouse::class;
 
-    /**
-     * Define the model's default state.
-     *
-     * @return array<string, mixed>
-     */
     public function definition(): array
     {
         return [
